abstract downloaders
