+++ downloader/s1_downloader.py
@@ -1,14 +1,8 @@
-from downloader.utils import load_json, process_dates, get_keycloak, download_file
-from downloader.xml_utils import parse_manifest, get_files
-from pathlib import Path
-from tqdm import tqdm
-import requests
-import time
+from downloader.base_downloader import SentinelDownloader
+from downloader.utils import load_json
+from pathlib import Path 
 
-data_url = "https://catalogue.dataspace.copernicus.eu/odata/v1"
-download_url = "https://download.dataspace.copernicus.eu/odata/v1"
-
-class Sentinel1:
+class Sentinel1(SentinelDownloader):
     def __init__(self, username, password, footprints_path, orbit_direction,
         product_type, polarization_mode, initial_date, last_date, output_dir):   
         """
@@ -17,23 +11,14 @@ class Sentinel1:
             orbit_direction (str): Orbit direction (can be "ASCENDING" or "DESCENDING").
             product_type (str): Sentinel-1 product type (e.g., "GRDH", "SLC").
             polarization_mode (list[str]): Polarization modes (e.g., ["VV", "VH"]).
-            initial_date (str): Start date to download, in "YYYY-MM-DD" format.
-            last_date (str): End date to download, in "YYYY-MM-DD" format.
-            output_dir (str): Path to save the files
         """   
+        super().__init__(username, password, initial_date, last_date, output_dir)
         self.data_collection = 'SENTINEL-1'
-        self.username = username
-        self.password = password
 
         self.orbit_direction = orbit_direction
         self.product_type = product_type
         self.polarization_mode = polarization_mode
 
-        self.initial_date = initial_date
-        self.last_date = last_date
-        self.date_ranges = process_dates(initial_date, last_date)
-
-        self.output_dir = Path(output_dir)
         self.footprints_path = Path(footprints_path)
         self.footprints = load_json(self.footprints_path)
 
@@ -65,39 +50,6 @@ class Sentinel1:
         description = ", ".join(attributes)
         return f"Sentinel-1({description})"
 
-    def init_session(self):
-        """
-        Initializes an authenticated session for accessing Copernicus Data Space API.
-
-        This method:
-        - Retrieves an access token using Keycloak authentication.
-        - Creates a new `requests.Session` instance.
-        - Updates the session headers to include the Bearer token for authorization.
-
-        Returns:
-            requests.Session: A configured session with authentication headers.
-        """
-        access_token = get_keycloak(self.username, self.password)
-        session = requests.Session()
-        session.headers.update({"Authorization": f"Bearer {access_token}"})
-        return session
-
-    def prepare_output(self, product_name):
-        """
-        Creates and returns the output directory for a given product.
-
-        This method ensures that the directory exists before returning its path.
-
-        Args:
-            product_name (str): Name of the product for which the output directory is created.
-
-        Returns:
-            pathlib.Path: Path to the created (or existing) output directory.
-        """
-        product_path = self.output_dir / product_name
-        product_path.mkdir(parents=True, exist_ok=True)
-        return product_path
-
     def get_query(self, tile_id, initial_date, last_date):
         """
         Constructs an OData query for retrieving Sentinel-1 products from the Copernicus Data Space API.
@@ -113,7 +65,7 @@ class Sentinel1:
         aoi = self.footprints[tile_id]
         footprint = ", ".join(aoi)
         query = [
-            f"{data_url}/Products?$filter=Collection/Name eq '{self.data_collection}'",
+            f"{self.data_url}/Products?$filter=Collection/Name eq '{self.data_collection}'",
             f"ContentDate/Start ge {initial_date}",
             f"ContentDate/End le {last_date}",
             f"OData.CSC.Intersects(area=geography'SRID=4326;POLYGON(({footprint}))')",
@@ -140,85 +92,6 @@ class Sentinel1:
               Download all the files for processing later.
         """
         return files_list
-
-    def download_product(self, session, SAFE_product):
-        """
-        Downloads a Sentinel-1 product and its relevant files.
-
-        This method:
-        1. Creates an output directory for the product.
-        2. Downloads the `manifest.safe` file if it does not already exist.
-        3. Parses the manifest file to retrieve the list of image files.
-        4. Filters image files based on predefined criteria.
-        5. Downloads the filtered image files if they do not already exist.
-
-        Args:
-            session (requests.Session): An authenticated session for making HTTP requests.
-            SAFE_product (dict): A dictionary containing product metadata with:
-                - "Id" (str): The product's unique identifier.
-                - "Name" (str): The product's name.
-
-        Notes:
-            - The function ensures that files are downloaded only if they are missing.
-            - The folder structure is preserved to match the Sentinel-1 SAFE format.
-        """
-        product_id = SAFE_product["Id"]
-        product_name = SAFE_product["Name"]
-        product_base_url = f"{download_url}/Products({product_id})/Nodes({product_name})"
-        
-        product_path = self.prepare_output(product_name)
-        
-        manifest_path = product_path / "manifest.safe"
-        if not manifest_path.is_file():
-            manifest_url = f"{product_base_url}/Nodes(manifest.safe)/$value"
-            response = session.get(manifest_url, allow_redirects=False)
-            download_file(response, manifest_path)
-            
-        xmldict = parse_manifest(manifest_path)
-        files_list = self.filter_images(get_files(xmldict))
-
-        for file in files_list:
-            file_path = product_path / file
-            file_path.parent.mkdir(parents=True, exist_ok=True)
-            if file_path.is_file():
-                continue 
-
-            nodes_str = "/".join([f"Nodes({node})" for node in file.split("/")])
-            file_url = f"{product_base_url}/{nodes_str}/$value"
-            response = session.get(file_url, allow_redirects=False)
-            download_file(response, file_path)
-
-    def download_tile(self, tile_id):
-        """
-        Downloads all Sentinel-1 products for a given AOI across multiple date ranges.
-
-        This method:
-        1. Iterates through predefined date ranges.
-        2. Constructs a query to fetch available Sentinel-1 products for the tile.
-        3. Sends a request to retrieve product metadata.
-        4. Iterates through each product, initializing a session and downloading the product files.
-        5. Introduces a delay between requests to prevent excessive API calls.
-
-        Args:
-            tile_id (str): The ID of an AOI of interest.
-
-        Notes:
-            - Uses `self.get_query()` to construct the API request URL.
-            - Uses `self.init_session()` to authenticate before downloading.
-            - Uses `self.download_product()` to handle the actual file downloads.
-            - Introduces a **10-second delay** (`time.sleep(10)`) between iterations to avoid rate limits.
-        """
-        for initial_date, last_date in self.date_ranges:
-            query = self.get_query(tile_id, initial_date, last_date)
-            response = requests.get(query)
-            data = response.json()
-
-            desc = f"Downloading tile {tile_id} from {initial_date[:10]} to {last_date[:10]}"
-            for SAFE_product in tqdm(data["value"], desc=desc):
-                print()
-                with self.init_session() as session:
-                    self.download_product(session, SAFE_product)
-            time.sleep(10)
 
     def download(self):
         """
